+++ micro-services/back-office-service/src/controllers/SalesController.ts
@@ -10,45 +10,24 @@ export class SalesController {
   }
 
   async getDashboardStats(req: Request, res: Response): Promise<void> {
-    try {
-      const fromRaw = this.normalizeQueryValue(req.query.from);
-      const toRaw = this.normalizeQueryValue(req.query.to);
+    const fromRaw = this.normalizeQueryValue(req.query.from);
+    const toRaw = this.normalizeQueryValue(req.query.to);
 
-      const fromDate = fromRaw ? new Date(fromRaw) : undefined;
-      const toDate = toRaw ? new Date(toRaw) : undefined;
+    const fromDate = fromRaw ? new Date(fromRaw) : undefined;
+    const toDate = toRaw ? new Date(toRaw) : undefined;
 
-      if (fromRaw && (!fromDate || Number.isNaN(fromDate.getTime()))) {
-        res.status(400).json({ success: false, message: 'Parametre "from" invalide.' });
-        return;
-      }
-
-      if (toRaw && (!toDate || Number.isNaN(toDate.getTime()))) {
-        res.status(400).json({ success: false, message: 'Parametre "to" invalide.' });
-        return;
-      }
-
-      if (fromDate && toDate && fromDate > toDate) {
-        res.status(400).json({ success: false, message: 'La date "from" doit etre inferieure a "to".' });
-        return;
-      }
-
-      const data = await this.service.getDashboardStats(fromDate, toDate);
-      res.status(200).json({ success: true, data });
-    } catch (err) {
-      const message = err instanceof Error ? err.message : 'Erreur interne.';
-      res.status(500).json({ success: false, message });
-    }
+    const data = await this.service.getDashboardStats(fromDate, toDate);
+    res.status(200).json({ success: true, data });
   }
 
+  // Normalise les query params Express qui peuvent être string | string[] | ParsedQs
   private normalizeQueryValue(value: unknown): string | undefined {
     if (Array.isArray(value)) {
-      return value[0];
+      return typeof value[0] === 'string' ? value[0].trim() || undefined : undefined;
     }
-
     if (typeof value === 'string') {
       return value.trim() || undefined;
     }
-
     return undefined;
   }
 }
